feat: file to house basic integration tests, and switch from using Click CLI runner to using subprocess
